refactor: move git utils out of commands folder
